+++ server/service/endpoint_scheduled_queries.go
@@ -8,83 +8,6 @@ import (
 )
 
 ////////////////////////////////////////////////////////////////////////////////
-// Get Scheduled Query
-////////////////////////////////////////////////////////////////////////////////
-
-type getScheduledQueryRequest struct {
-	ID uint
-}
-
-type getScheduledQueryResponse struct {
-	Scheduled *scheduledQueryResponse `json:"scheduled,omitempty"`
-	Err       error                   `json:"error,omitempty"`
-}
-
-func (r getScheduledQueryResponse) error() error { return r.Err }
-
-func makeGetScheduledQueryEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(getScheduledQueryRequest)
-
-		sq, err := svc.GetScheduledQuery(ctx, req.ID)
-		if err != nil {
-			return getScheduledQueryResponse{Err: err}, nil
-		}
-
-		return getScheduledQueryResponse{
-			Scheduled: &scheduledQueryResponse{
-				ScheduledQuery: *sq,
-			},
-		}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Schedule Query
-////////////////////////////////////////////////////////////////////////////////
-
-type scheduleQueryRequest struct {
-	PackID   uint    `json:"pack_id"`
-	QueryID  uint    `json:"query_id"`
-	Interval uint    `json:"interval"`
-	Snapshot *bool   `json:"snapshot"`
-	Removed  *bool   `json:"removed"`
-	Platform *string `json:"platform"`
-	Version  *string `json:"version"`
-	Shard    *uint   `json:"shard"`
-}
-
-type scheduleQueryResponse struct {
-	Scheduled *scheduledQueryResponse `json:"scheduled,omitempty"`
-	Err       error                   `json:"error,omitempty"`
-}
-
-func (r scheduleQueryResponse) error() error { return r.Err }
-
-func makeScheduleQueryEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(scheduleQueryRequest)
-
-		scheduled, err := svc.ScheduleQuery(ctx, &fleet.ScheduledQuery{
-			PackID:   req.PackID,
-			QueryID:  req.QueryID,
-			Interval: req.Interval,
-			Snapshot: req.Snapshot,
-			Removed:  req.Removed,
-			Platform: req.Platform,
-			Version:  req.Version,
-			Shard:    req.Shard,
-		})
-		if err != nil {
-			return scheduleQueryResponse{Err: err}, nil
-		}
-		return scheduleQueryResponse{Scheduled: &scheduledQueryResponse{
-			ScheduledQuery: *scheduled,
-		}}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
 // Modify Scheduled Query
 ////////////////////////////////////////////////////////////////////////////////
